Added support for float or double osg::Plane, and the default osg::Plane to double. Performance tests on big models did not indicate any performance penalty in using doubles over floats, so the move to doubles should mainly impact precision improvements for whole earth databases.
Also made improvements to osgUtil::PlaneIntersector and osgSim::ElevationSlice classes

+++ src/osgPlugins/ive/DataOutputStream.cpp
@@ -272,10 +272,10 @@ void DataOutputStream::writeVec4d(const osg::Vec4d& v){
 
 void DataOutputStream::writePlane(const osg::Plane& v)
 {
-    writeFloat(v[0]);
-    writeFloat(v[1]);
-    writeFloat(v[2]);
-    writeFloat(v[3]);
+    writeDouble(v[0]);
+    writeDouble(v[1]);
+    writeDouble(v[2]);
+    writeDouble(v[3]);
     
     if (_verboseOutput) std::cout<<"read/writePlane() ["<<v<<"]"<<std::endl;
 }
